extract argfactory
